set up image saver class and header
